# docs/openapi.json
{"openapi": "3.1.0", "info": {"title": "Nifty 100 Financial API", "version": "1.0.0"}, "paths": {"/api/v1/health": {"get": {"tags": ["Health"], "summary": "Health Check", "operationId": "health_check_api_v1_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/": {"get": {"tags": ["Companies"], "summary": "Get All Companies", "operationId": "get_all_companies_api_v1__get", "parameters": [{"name": "sector", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sector"}}, {"name": "search", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Search"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{ticker}": {"get": {"tags": ["Companies"], "summary": "Get Company Profile", "operationId": "get_company_profile_api_v1__ticker__get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{ticker}/pl": {"get": {"tags": ["Companies"], "summary": "Get Profit And Loss", "operationId": "get_profit_and_loss_api_v1__ticker__pl_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{ticker}/bs": {"get": {"tags": ["Companies"], "summary": "Get Balance Sheet", "operationId": "get_balance_sheet_api_v1__ticker__bs_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{ticker}/cashflow": {"get": {"tags": ["Companies"], "summary": "Get Cashflow", "operationId": "get_cashflow_api_v1__ticker__cashflow_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{ticker}/ratios": {"get": {"tags": ["Companies"], "summary": "Get Ratios", "operationId": "get_ratios_api_v1__ticker__ratios_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{ticker}/tearsheet": {"get": {"tags": ["Companies"], "summary": "Download Tearsheet", "operationId": "download_tearsheet_api_v1__ticker__tearsheet_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/screener": {"get": {"tags": ["Screener"], "summary": "Run Screener", "operationId": "run_screener_api_v1_screener_get", "parameters": [{"name": "min_roe", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Min Roe"}}, {"name": "max_de", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Max De"}}, {"name": "min_fcf", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Min Fcf"}}, {"name": "sector", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sector"}}, {"name": "min_rev_cagr_5yr", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Min Rev Cagr 5Yr"}}, {"name": "min_pat_cagr_5yr", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Min Pat Cagr 5Yr"}}, {"name": "max_pe", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Max Pe"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/sectors": {"get": {"tags": ["Sectors"], "summary": "Get Sectors", "operationId": "get_sectors_api_v1_sectors_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/sectors/{sector}/companies": {"get": {"tags": ["Sectors"], "summary": "Get Sector Companies", "operationId": "get_sector_companies_api_v1_sectors__sector__companies_get", "parameters": [{"name": "sector", "in": "path", "required": true, "schema": {"type": "string", "title": "Sector"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/peers/{group_name}": {"get": {"tags": ["Peers"], "summary": "Get Peer Group", "operationId": "get_peer_group_api_v1_peers__group_name__get", "parameters": [{"name": "group_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Group Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/peers/compare": {"get": {"tags": ["Peers"], "summary": "Get Peer Comparison", "operationId": "get_peer_comparison_api_v1_companies__ticker__peers_compare_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/market-cap/{ticker}": {"get": {"tags": ["Valuation"], "summary": "Get Market Cap", "operationId": "get_market_cap_api_v1_market_cap__ticker__get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/portfolio/stats": {"get": {"tags": ["Portfolio"], "summary": "Get Portfolio Stats", "operationId": "get_portfolio_stats_api_v1_portfolio_stats_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/companies/{ticker}/documents": {"get": {"tags": ["Documents"], "summary": "Get Documents", "operationId": "get_documents_api_v1_companies__ticker__documents_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}